+++ packages/gatsby-transformer-javascript-static-exports/src/gatsby-node.js
@@ -12,7 +12,7 @@ async function onCreateNode({
 }) {
   const { createNode, createParentChildLink } = actions
 
-  // This only processes javascript files.
+  // This only processes JavaScript files.
   if (node.internal.mediaType !== `application/javascript`) {
     return
   }