+++ packages/gatsby/cache-dir/ensure-resources.js
@@ -2,6 +2,7 @@ import React from "react"
 import PropTypes from "prop-types"
 import loader from "./loader"
 import shallowCompare from "shallow-compare"
+import { getRedirectUrl } from "./load-directly-or-404"
 
 // Pass pathname in as prop.
 // component will try fetching resources. If they exist,
@@ -97,15 +98,11 @@ class EnsureResources extends React.Component {
       process.env.NODE_ENV === `production` &&
       !(this.state.pageResources && this.state.pageResources.json)
     ) {
-      // Try to load the page directly - this should result in a 404 or
-      // network offline error.
-      const url = new URL(window.location)
-      if (url.search) {
-        url.search += `&no-cache=1`
-      } else {
-        url.search = `?no-cache=1`
+      // This should only occur if there's no custom 404 page
+      const url = getRedirectUrl(this.state.location.href)
+      if (url) {
+        window.location.replace(url)
       }
-      window.location.replace(url)
 
       return null
     }