+++ packages/gatsby/cache-dir/json-store.js
@@ -90,11 +90,7 @@ class JSONStore extends React.Component {
 
     return (
       <StaticQueryContext.Provider value={this.state.staticQueryData}>
-        <PageRenderer
-          key={this.props.location.pathname}
-          {...propsWithoutPages}
-          {...data}
-        />
+        <PageRenderer {...propsWithoutPages} {...data} />
       </StaticQueryContext.Provider>
     )
   }