+++ packages/gatsby/cache-dir/static-entry.js
@@ -52,7 +52,7 @@ const createElement = React.createElement
 export default (pagePath, callback) => {
   let bodyHtml = ``
   let headComponents = [
-    <meta name="generator" content={`Gatsby ${gatsbyVersion}`} />
+    <meta name="generator" content={`Gatsby ${gatsbyVersion}`} />,
   ]
   let htmlAttributes = {}
   let bodyAttributes = {}
@@ -163,10 +163,10 @@ export default (pagePath, callback) => {
 
   const bodyComponent = apiRunner(
     `wrapRootElement`,
-    { element: routerElement },
+    { element: routerElement, pathname: pagePath },
     routerElement,
     ({ result }) => {
-      return { element: result }
+      return { element: result, pathname: pagePath }
     }
   ).pop()
 
@@ -180,6 +180,8 @@ export default (pagePath, callback) => {
     setPreBodyComponents,
     setPostBodyComponents,
     setBodyProps,
+    pathname: pagePath,
+    pathPrefix: __PATH_PREFIX__,
   })
 
   // If no one stepped up, we'll handle it.
@@ -315,15 +317,6 @@ export default (pagePath, callback) => {
       }
     })
 
-  apiRunner(`onPreRenderHTML`, {
-    getHeadComponents,
-    replaceHeadComponents,
-    getPreBodyComponents,
-    replacePreBodyComponents,
-    getPostBodyComponents,
-    replacePostBodyComponents,
-  })
-
   // Add page metadata for the current page
   const windowData = `/*<![CDATA[*/window.page=${JSON.stringify(page)};${
     page.jsonName in dataPaths
@@ -367,6 +360,17 @@ export default (pagePath, callback) => {
   })
 
   postBodyComponents.push(...bodyScripts)
+
+  apiRunner(`onPreRenderHTML`, {
+    getHeadComponents,
+    replaceHeadComponents,
+    getPreBodyComponents,
+    replacePreBodyComponents,
+    getPostBodyComponents,
+    replacePostBodyComponents,
+    pathname: pagePath,
+    pathPrefix: __PATH_PREFIX__,
+  })
 
   const html = `<!DOCTYPE html>${renderToStaticMarkup(
     <Html