+++ www/src/components/plugin-searchbar-body.js
@@ -18,7 +18,8 @@ import unescape from "lodash/unescape"
 import presets from "../utils/presets"
 import typography, { rhythm, scale } from "../utils/typography"
 import { scrollbarStyles } from "../utils/styles"
-import { css as glam } from "glamor"
+import { injectGlobal } from "react-emotion"
+import removeMD from "remove-markdown"
 
 // This is for the urlSync
 const updateAfter = 700
@@ -28,7 +29,8 @@ const searchInputHeight = rhythm(7 / 4)
 const searchMetaHeight = rhythm(8 / 4)
 const searchInputWrapperMargin = rhythm(3 / 4)
 
-glam.insert(`
+/* stylelint-disable */
+injectGlobal`
   .ais-SearchBox__input:valid ~ .ais-SearchBox__reset {
     display: block;
   }
@@ -167,7 +169,8 @@ glam.insert(`
   .ais-InfiniteHits__loadMore[disabled] {
     display: none;
   }
-`)
+`
+/* stylelint-enable */
 
 // Search shows a list of "hits", and is a child of the PluginSearchBar component
 class Search extends Component {
@@ -330,7 +333,7 @@ const Result = ({ hit, pathname, search }) => {
           "&:before": {
             background: colors.ui.border,
             bottom: 0,
-            content: ` `,
+            content: `''`,
             height: 1,
             left: 0,
             position: `absolute`,
@@ -340,7 +343,7 @@ const Result = ({ hit, pathname, search }) => {
           "&:after": {
             background: selected ? colors.gatsby : false,
             bottom: 0,
-            content: ` `,
+            content: `''`,
             position: `absolute`,
             left: 0,
             top: -1,
@@ -396,7 +399,7 @@ const Result = ({ hit, pathname, search }) => {
           lineHeight: 1.5,
         }}
       >
-        {unescape(hit.description)}
+        {removeMD(unescape(hit.description))}
       </div>
     </Link>
   )