+++ www/src/templates/template-blog-list.js
@@ -18,7 +18,8 @@ class BlogPostsIndex extends React.Component {
 
     return (
       <Layout location={this.props.location}>
-        <div
+        <main
+          id={`reach-skip-nav`}
           css={{
             [presets.Tablet]: {
               background: colors.ui.whisper,
@@ -100,7 +101,7 @@ class BlogPostsIndex extends React.Component {
             <Pagination context={this.props.pageContext} />
             <EmailCaptureForm signupMessage="Enjoying our blog? Receive the next post in your inbox!" />
           </Container>
-        </div>
+        </main>
       </Layout>
     )
   }
@@ -115,6 +116,7 @@ export const pageQuery = graphql`
       filter: {
         frontmatter: { draft: { ne: true } }
         fileAbsolutePath: { regex: "/docs.blog/" }
+        fields: { released: { eq: true } }
       }
       limit: $limit
       skip: $skip