+++ www/src/views/showcase/showcase-list.js
@@ -3,6 +3,7 @@ import { Link } from "gatsby"
 
 import styles from "../shared/styles"
 import ThumbnailLink from "../shared/thumbnail"
+import EmptyGridItems from "../shared/empty-grid-items"
 import qs from "qs"
 
 import ShowcaseItemCategories from "./showcase-item-categories"
@@ -16,11 +17,7 @@ const ShowcaseList = ({ items, count }) => {
   if (count) items = items.slice(0, count)
 
   return (
-    <div
-      css={{
-        ...styles.showcaseList,
-      }}
-    >
+    <main id={`reach-skip-nav`} css={{ ...styles.showcaseList }}>
       {items.map(
         ({ node }) =>
           node.fields &&
@@ -79,11 +76,7 @@ const ShowcaseList = ({ items, count }) => {
                 </div>
                 {node.featured && (
                   <Link
-                    css={{
-                      "&&": {
-                        ...styles.featuredItem,
-                      },
-                    }}
+                    css={{ "&&": { ...styles.featuredItem } }}
                     to={`/showcase?${qs.stringify({
                       filters: `Featured`,
                     })}`}
@@ -92,9 +85,7 @@ const ShowcaseList = ({ items, count }) => {
                     <img
                       src={FeaturedIcon}
                       alt="icon"
-                      css={{
-                        ...styles.featuredIcon,
-                      }}
+                      css={{ ...styles.featuredIcon }}
                     />
                   </Link>
                 )}
@@ -102,32 +93,8 @@ const ShowcaseList = ({ items, count }) => {
             </div>
           )
       )}
-      {/* makes last row items equal width and aligned left */}
-      <div
-        aria-hidden="true"
-        css={{ ...styles.showcaseItem, marginTop: 0, marginBottom: 0 }}
-      />
-      <div
-        aria-hidden="true"
-        css={{ ...styles.showcaseItem, marginTop: 0, marginBottom: 0 }}
-      />
-      <div
-        aria-hidden="true"
-        css={{ ...styles.showcaseItem, marginTop: 0, marginBottom: 0 }}
-      />
-      <div
-        aria-hidden="true"
-        css={{ ...styles.showcaseItem, marginTop: 0, marginBottom: 0 }}
-      />
-      <div
-        aria-hidden="true"
-        css={{ ...styles.showcaseItem, marginTop: 0, marginBottom: 0 }}
-      />
-      <div
-        aria-hidden="true"
-        css={{ ...styles.showcaseItem, marginTop: 0, marginBottom: 0 }}
-      />
-    </div>
+      {items.length && <EmptyGridItems styles={styles.showcaseItem} />}
+    </main>
   )
 }