+++ packages/gatsby-cli/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.4.5"></a>
+
+## [2.4.5](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.4.4...gatsby-cli@2.4.5) (2018-11-08)
+
+**Note:** Version bump only for package gatsby-cli
+
+<a name="2.4.4"></a>
+
+## [2.4.4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.4.3...gatsby-cli@2.4.4) (2018-10-29)
+
+**Note:** Version bump only for package gatsby-cli
+
+<a name="2.4.3"></a>
+
+## [2.4.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.4.2...gatsby-cli@2.4.3) (2018-10-09)
+
+### Features
+
+- add error message with filename on Markdown error, fix bug in panicOnBuild ([#8866](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/issues/8866)) ([bbff3be](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/commit/bbff3be))
+
+<a name="2.4.2"></a>
+
+## [2.4.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.4.1...gatsby-cli@2.4.2) (2018-09-27)
+
+### Bug Fixes
+
+- add compat fix for gatsby-cli v2 with gatsby v1 ([#8581](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/issues/8581)) ([279ea76](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/commit/279ea76))
+
+<a name="2.4.1"></a>
+
+## [2.4.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.4.0...gatsby-cli@2.4.1) (2018-09-18)
+
+### Features
+
+- add --prefix-paths option to gatsby serve cli ([#8060](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/issues/8060)) ([98c8e91](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/commit/98c8e91))
+
 <a name="2.4.0"></a>
 
 # [2.4.0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.0.0-rc.6...gatsby-cli@2.4.0) (2018-09-17)