+++ packages/gatsby-plugin-manifest/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.8"></a>
+
+## [2.0.8](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-manifest/compare/gatsby-plugin-manifest@2.0.7...gatsby-plugin-manifest@2.0.8) (2018-11-08)
+
+**Note:** Version bump only for package gatsby-plugin-manifest
+
+<a name="2.0.7"></a>
+
+## [2.0.7](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-manifest/compare/gatsby-plugin-manifest@2.0.6...gatsby-plugin-manifest@2.0.7) (2018-10-29)
+
+**Note:** Version bump only for package gatsby-plugin-manifest
+
+<a name="2.0.6"></a>
+
+## [2.0.6](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-manifest/compare/gatsby-plugin-manifest@2.0.5...gatsby-plugin-manifest@2.0.6) (2018-10-24)
+
+**Note:** Version bump only for package gatsby-plugin-manifest
+
+<a name="2.0.5"></a>
+
+## [2.0.5](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-manifest/compare/gatsby-plugin-manifest@2.0.4...gatsby-plugin-manifest@2.0.5) (2018-10-09)
+
+**Note:** Version bump only for package gatsby-plugin-manifest
+
+<a name="2.0.4"></a>
+
+## [2.0.4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-manifest/compare/gatsby-plugin-manifest@2.0.3...gatsby-plugin-manifest@2.0.4) (2018-09-26)
+
+**Note:** Version bump only for package gatsby-plugin-manifest
+
+<a name="2.0.3"></a>
+
+## [2.0.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-manifest/compare/gatsby-plugin-manifest@2.0.2-rc.1...gatsby-plugin-manifest@2.0.3) (2018-09-24)
+
+### Bug Fixes
+
+- **gatsby-plugin-manifest:** favicon path respects hybrid mode ([#8315](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-manifest/issues/8315)) ([05e7ccd](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-manifest/commit/05e7ccd))
+
 <a name="2.0.2"></a>
 
 ## [2.0.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-manifest/compare/gatsby-plugin-manifest@2.0.2-rc.1...gatsby-plugin-manifest@2.0.2) (2018-09-17)