+++ packages/gatsby-plugin-offline/CHANGELOG.md
@@ -3,6 +3,64 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.13"></a>
+
+## [2.0.13](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-offline/compare/gatsby-plugin-offline@2.0.12...gatsby-plugin-offline@2.0.13) (2018-11-08)
+
+**Note:** Version bump only for package gatsby-plugin-offline
+
+<a name="2.0.12"></a>
+
+## [2.0.12](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-offline/compare/gatsby-plugin-offline@2.0.11...gatsby-plugin-offline@2.0.12) (2018-11-05)
+
+### Bug Fixes
+
+- **gatsby-plugin-offline:** Serve the offline shell for short URLs + use no-cors for external resources ([#9679](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-offline/issues/9679)) ([430e8f1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-offline/commit/430e8f1)), closes [#8145](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-offline/issues/8145)
+
+<a name="2.0.11"></a>
+
+## [2.0.11](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-offline/compare/gatsby-plugin-offline@2.0.10...gatsby-plugin-offline@2.0.11) (2018-11-01)
+
+### Bug Fixes
+
+- **gatsby-plugin-offline:** don't precache the index page ([#9603](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-offline/issues/9603)) ([00284e0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-offline/commit/00284e0)), closes [#7997](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-offline/issues/7997)
+
+<a name="2.0.10"></a>
+
+## [2.0.10](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-offline/compare/gatsby-plugin-offline@2.0.9...gatsby-plugin-offline@2.0.10) (2018-10-29)
+
+**Note:** Version bump only for package gatsby-plugin-offline
+
+<a name="2.0.9"></a>
+
+## [2.0.9](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-offline/compare/gatsby-plugin-offline@2.0.8...gatsby-plugin-offline@2.0.9) (2018-10-24)
+
+**Note:** Version bump only for package gatsby-plugin-offline
+
+<a name="2.0.8"></a>
+
+## [2.0.8](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-offline/compare/gatsby-plugin-offline@2.0.7...gatsby-plugin-offline@2.0.8) (2018-10-23)
+
+### Features
+
+- update Workbox to 3.6.3 ([#9294](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-offline/issues/9294)) ([f53d457](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-offline/commit/f53d457))
+
+<a name="2.0.7"></a>
+
+## [2.0.7](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-offline/compare/gatsby-plugin-offline@2.0.6...gatsby-plugin-offline@2.0.7) (2018-10-16)
+
+### Bug Fixes
+
+- update gatsby peerDep version ([#9150](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-offline/issues/9150)) ([f5c5556](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-offline/commit/f5c5556))
+
+<a name="2.0.6"></a>
+
+## [2.0.6](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-offline/compare/gatsby-plugin-offline@2.0.5...gatsby-plugin-offline@2.0.6) (2018-10-09)
+
+### Bug Fixes
+
+- **gatsby-plugin-offline:** delay adding resources for paths until we have urls ([#8613](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-offline/issues/8613)) ([2605aa0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-offline/commit/2605aa0))
+
 <a name="2.0.5"></a>
 
 ## [2.0.5](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-offline/compare/gatsby-plugin-offline@2.0.0-rc.9...gatsby-plugin-offline@2.0.5) (2018-09-17)